make more stable after Cluster Failover test
* with some refactoring
